[gcc] MenuPublic: Don't round master volume setting

+++ src/gui/MenuPublic.h
@@ -57,7 +57,7 @@ void ARXMenu_Options_Video_SetDetailsQuality(int lod);
 // Audio options
 
 void ARXMenu_Options_Audio_SetDevice(const std::string & device);
-void ARXMenu_Options_Audio_SetMasterVolume(int volume);
+void ARXMenu_Options_Audio_SetMasterVolume(float volume);
 void ARXMenu_Options_Audio_SetSfxVolume(int volume);
 void ARXMenu_Options_Audio_SetSpeechVolume(int volume);
 void ARXMenu_Options_Audio_SetAmbianceVolume(int volume);
